refactor: Full note
